process_data/requirements.txt
chromadb==0.5.20
contractions==0.1.73
polars-lts-cpu==1.18.0
pydantic==2.10.2
nltk==3.9.1
llama-index-embeddings-huggingface==0.5.1